services/kafka/requirements.txt

requests
confluent-kafka
psycopg2-binary
redis
google-genai
